# apis/product-service/swagger.yml
openapi: 3.0.0
info:
  title: Customer API
  description: API for managing customers
  version: 1.0.0
paths:
  "/v1/customers":
    get:
      summary: Get all customers
      tags:
        - CustomersV1
      responses:
        '200':
          description: A list of customers
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                  elapsed_time:
                    type: string
                  data:
                    type: object
                    properties:
                      page_size:
                        type: integer
                      page_content:
                        type: array
                        items:
                          $ref: '#/components/schemas/Customer'
    post:
      tags:
        - CustomersV1
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerWrite'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                  elapsed_time:
                    type: string
                  data:
                    type: object
                    properties:
                      id:
                        type: string

  "/v1/customers/{id}":
    get:
      tags:
        - CustomersV1
      summary: Get a customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Customer details
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                  elapsed_time:
                    type: string
                  data:
                    type: object
                    properties:
                      customer:
                        $ref: '#/components/schemas/Customer'
    put:
      tags:
        - CustomersV1
      summary: Update a customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerWrite'
      responses:
        '200':
          description: Customer updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                  elapsed_time:
                    type: string
                  data:
                    type: object
                    properties:
                      id:
                        type: string
    delete:
      tags:
        - CustomersV1
      summary: Delete a customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Customer deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                  elapsed_time:
                    type: string

components:
  schemas:
    Customer:
      type: object
      properties:
        customer_id:
          type: string
          example: "01HZ7E8GR7SBPV9F96XRR5HCW2"
        name:
          type: string
          example: "Juninho"
        surname:
          type: string
          example: "Doença"
        email:
          type: string
          example: "doenca@gmail.com"
        birthdate:
          type: string
          format: date
          example: "1990-03-07"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
    CustomerWrite:
      type: object
      properties:
        name:
          type: string
          example: "Juninho"
        surname:
          type: string
          example: "Doença"
        email:
          type: string
          example: "doenca@gmail.com"
        birthdate:
          type: string
          format: date
          example: "1990-03-07"